requirements.txt
bitbucket-webhooks==0.0.10
Flask==1.1.2
uWSGI==2.0.19.1
python-dotenv
